Enable bugprone widening checks in clang
Most of the errors we hit are simply places we need to explicitly
increase the width of the integer.  Luckily, these are few and far
between.

Tested: Code compiles, unit tests pass.

+++ http/http_connection.hpp
@@ -38,8 +38,8 @@ inline void prettyPrintJson(crow::Response& res)
 static int connectionCount = 0;
 
 // request body limit size set by the bmcwebHttpReqBodyLimitMb option
-constexpr unsigned int httpReqBodyLimit =
-    1024 * 1024 * bmcwebHttpReqBodyLimitMb;
+constexpr uint64_t httpReqBodyLimit =
+    1024UL * 1024UL * bmcwebHttpReqBodyLimitMb;
 
 constexpr uint64_t loggedOutPostBodyLimit = 4096;
